+++ tests/fixtures.py
@@ -1,5 +1,7 @@
+import os
 import pytest
-from lifxlan import LifxLAN, MultiZoneLight
+from config import config
+from lifxlan import MultiZoneLight
 
 
 def lifx_lights():
@@ -23,3 +25,16 @@ def lifx_lights():
 @pytest.fixture(scope='function')
 def get_lights():
     return lifx_lights()
+
+
+@pytest.fixture(scope='function')
+def setup_light_store():
+    lightstore_dir = os.path.join(config.DIR, 'tests/lightstore')
+    if not os.path.exists(lightstore_dir):
+        os.mkdir(lightstore_dir)
+    return lightstore_dir
+
+
+@pytest.fixture
+def light_store(setup_light_store):
+    return os.path.join(config.DIR, 'tests/lightstore')
