+++ tests/test_lights.py
@@ -1,31 +1,14 @@
 import os
 import pickle
 import shutil
-
-import pytest
 from unittest import mock
 
 from tests.fixtures import *
 from lifxlan import LifxLAN, MultiZoneLight
-
-from config import config
 from backend import lights
 
 lan = LifxLAN()
 m = mock.Mock()
-
-
-@pytest.fixture(scope='function')
-def setup_light_store():
-    lightstore_dir = os.path.join(config.DIR, 'tests/lightstore')
-    if not os.path.exists(lightstore_dir):
-        os.mkdir(lightstore_dir)
-    return lightstore_dir
-
-
-@pytest.fixture
-def light_store(setup_light_store):
-    return os.path.join(config.DIR, 'tests/lightstore')
 
 
 def clean_up_light_store(light_store):
@@ -48,3 +31,12 @@ def test_store_light(get_lights, setup_light_store):
     clean_up_light_store(light_store)
 
 
+# @mock.patch('lan.get_lights', return_value=get_lights)
+def x_test_get_or_create_light(mocker):
+    stub = mocker.stub(name='get_lights_stub')
+
+    lan.get_lights(stub)
+    # mocker.patch('lan.get_lights', get_lights)
+    light_obj = lights.get_or_create_light("Standing Light", "12:34:56:78:9a:bc")
+    print('getting light object:', light_obj)
+    assert False
